Changes for latest core
