+++ devices/AliceSatellite.py
@@ -1,0 +1,33 @@
+import sqlite3
+from pathlib import Path
+from typing import Dict, Union
+
+from core.device.model.Device import Device
+from core.device.model.DeviceAbility import DeviceAbility
+
+
+class AliceSatellite(Device):
+
+	@classmethod
+	def getDeviceTypeDefinition(cls) -> dict:
+		return {
+			'deviceTypeName'        : 'AliceSatellite',
+			'perLocationLimit'      : 1,
+			'totalDeviceLimit'      : 0,
+			'allowLocationLinks'    : True,
+			'allowHeartbeatOverride': True,
+			'heartbeatRate'         : 5,
+			'deviceSettings'        : dict(),
+			'abilities'             : [DeviceAbility.PLAY_SOUND, DeviceAbility.CAPTURE_SOUND]
+		}
+
+
+	def __init__(self, data: Union[sqlite3.Row, Dict]):
+		super().__init__(data)
+
+
+	def getDeviceIcon(self) -> Path:
+		if not self.connected:
+			return Path(f'{self.Commons.rootDir()}/skills/{self.skillName}/device/img/{self.deviceTypeName}.png')
+		elif self.connected:
+			return Path(f'{self.Commons.rootDir()}/skills/{self.skillName}/device/img/connected.png')
