+++ devices/img/connected.png
