+++ devices/img/satellite_muted.png
