+++ devices/img/satellite_offline.png
